+++ api/README.md
@@ -144,3 +144,17 @@ Nombre de funcion = GET_FeedbackPayment.
     Esta es la ruta a la que deriva mercadopago al finalizar el pago, ya sea que se realizo exitosamente, como si fallo el pago.
     El request se realiza por body
 ```
+
+#### GET -> localhost:3700/orders
+
+```env
+Nombre de funcion = GET_AllOrders.
+    Se obtienen todas las ordenes.
+```
+
+#### GET -> localhost:3700/orders/:id
+
+```env
+Nombre de funcion = GET_OrderById.
+    Si se recibe por param el id, busca la orden relacionada a ese id
+```
