+++ api/src/Orders/controllers/controllers_order.ts
@@ -1,11 +1,14 @@
 import { Express, Request, Response } from "express";
 import { Op } from "sequelize";
 import db from "../../database";
+import dotenv from "dotenv";
 const mercadopago = require('mercadopago');
+dotenv.config();
 
 //MERCADOPAGO
 mercadopago.configure({
   access_token: process.env.MERCADOPAGO_KEY,
+  //access_token: "APP_USR-4964430421416242-020813-c46f247ea7b1f91937c722b8ea7b4134-1305644016",
 });
 
 interface RequestParams {}
@@ -13,8 +16,8 @@ interface RequestParams {}
 interface ResponseBody {}
 
 interface RequestBody {
-  id: number;     //id del producto
-  title: string;  //Este es el name del producto
+  id: number;         //id del producto
+  title: string;      //Este es el name del producto
   price: number;
   quantity: number;
   area_code: number;  //Telefono
@@ -85,7 +88,7 @@ export const POST_GeneratePayment =async (
 		});
 }
 
-//ruta de respuesta cuando el pago se realiza correctamente y cuando falla
+//ruta de respuesta de mercadopago cuando el pago se realiza exitosamente y cuando falla
 export const GET_FeedbackPayment =async (
   request: Request,
   response: Response
@@ -100,3 +103,54 @@ export const GET_FeedbackPayment =async (
 	});
 }
 
+//Obtener todas las ordenes
+export const GET_AllOrders = async (req: Request, res: Response) => {
+  try {
+      const orders = await db.Orders.findAll({
+          include: [
+              {
+                  model: db.Users,
+                  as: "users"
+              }
+          ]
+      });
+      return res.status(200).json(orders);
+    } catch (error: any) {
+      return res.status(400).json({message: error.message});
+    }
+  };
+
+// Obtener oden por ID
+export const GET_OrderById = async (req: Request, res: Response) => {
+  try {
+      const {id} = req.params;
+      const order = await db.Orders.findOne({
+          where: {id},
+          include: [
+            {
+                  model: db.Users,
+                  as: "user"
+                }
+              ]
+            });        if (!order) {
+              return res.status(404).json({message: "Orden no encontrada"});
+            }
+            return res.status(200).json(order);
+          } catch (error: any) {
+            return res.status(400).json({
+              message: error.message
+            });
+          }
+        };
+        
+        
+// Create a new order
+export const createOrder = async (req: Request, res: Response) => {
+  try {
+      const order = req.body;
+      const newOrder = await db.Orders.create(order);
+      return res.status(201).json(newOrder);
+  } catch (error: any) {
+      return res.status(400).json({message: error.message});
+  }
+};
