+++ api/src/Orders/routes/index.ts
@@ -3,13 +3,21 @@ const ordersRoutes = Router();
 
 import {
     POST_GeneratePayment,
-    GET_FeedbackPayment    
+    GET_FeedbackPayment,
+    GET_AllOrders,
+    GET_OrderById  
 } from "../controllers/controllers_order";
 
 //MERCADOPAGO
 ordersRoutes.post("/payment/", POST_GeneratePayment);
 
 ordersRoutes.get("/feedback/", GET_FeedbackPayment);
+
+//GET todas las ordenes
+ordersRoutes.get("/", GET_AllOrders);
+
+//GET orden por ID
+ordersRoutes.get("/:id", GET_OrderById);
 
 export default ordersRoutes;
